+++ editor/src/main/java/com/kotcrab/vis/editor/module/editor/FileChooserModule.java
@@ -39,6 +39,7 @@ public class FileChooserModule extends EditorModule {
 	@Override
 	public void init () {
 		FileChooser.setDefaultPrefsName("com.kotcrab.vis.editor");
+		FileChooser.setSaveLastDirectory(true);
 		chooser = new FileChooser(Mode.OPEN);
 		chooser.setSelectionMode(SelectionMode.FILES_AND_DIRECTORIES);
 		chooser.setFileDeleter(new JNAFileDeleter());