+++ ui/src/test/java/com/kotcrab/vis/ui/test/manual/TestFileChooser.java
@@ -34,6 +34,7 @@ public class TestFileChooser extends VisWindow {
 		super("filechooser");
 
 		FileChooser.setDefaultPrefsName("com.kotcrab.vis.ui.test.manual");
+		FileChooser.setSaveLastDirectory(true);
 		final FileChooser chooser = new FileChooser(Mode.OPEN);
 		chooser.setSelectionMode(FileChooser.SelectionMode.FILES_AND_DIRECTORIES);
 		chooser.setMultiSelectionEnabled(true);